Remove superfluous layers of indirection
